Add create method to Column class
The create method has been added to the Column class in the src/Column.php file. This static factory method simplifies the process of creating new column instances by not requiring the 'new' keyword to instantiate the object. It allows for flexible instantiation with custom parameters for name, column type, and size.